repackage Qt installer script code.
The separation of the bash shell code and the Qt Installer script
code allows the later to be used in environments were bash is not
available.  This was implemented with key,value pairs passed on
the command line, but it could have been done with environmental
variables using the installer environmentVariable method instead
of the value method.

add Qt 5.12.5 to regression.

add windows test cases to regression.  You can see how the installer script
code can be passed directly to the installer.  Even though the travis test case uses bash
in the yml file the separation of the script allows the installer and script to be used when bash isn't available.

add support for installs on windows, which have a unique page, see Controller.prototype.StartMenuDirectoryPageCallback

add support for new usage stats page, see Controller.prototype.DynamicTelemetryPluginFormCallback.
Recent installs with the online installer would fail due to this new page.
See https://www.qt.io/blog/option-to-provide-anonymous-usage-statistics-enabled.

Delete unused function in Qt installer script (abortInstaller).